fix(utils): Add if statement in bits_counter function in order to handle the null value. Therefore, update test function.

+++ include/utils/binary.h
@@ -12,6 +12,6 @@ int add_bits_to_buffer(uint8_t* buffer, const uint8_t* content,
                        const size_t content_len, int* bit_pos,
                        size_t buffer_byte_len);
 
-int bits_counter(int value);
+int bits_counter(uint8_t value);
 
 #endif  // _BINARY_H_
